+++ src/main/java/com/ae2subnet/api/ISubnetWorker.java
@@ -1,10 +1,11 @@
 package com.ae2subnet.api;
 
 import appeng.api.crafting.IPatternDetails;
+import appeng.api.networking.IGridNodeService;
 import appeng.api.stacks.KeyCounter;
 import net.minecraft.core.BlockPos;
 
-public interface ISubnetWorker {
+public interface ISubnetWorker extends IGridNodeService {
     WorkerState getWorkerState();
 
     boolean canAcceptInputs(KeyCounter[] inputHolder);
@@ -18,4 +19,8 @@ public interface ISubnetWorker {
     BlockPos getWorkerPos();
 
     boolean isValidWorker();
+
+    default boolean isRemoved() {
+        return false;
+    }
 }